column selection clean up
