+++ cpp/benchmarks/io/cuio_benchmark_common.hpp
@@ -24,10 +24,6 @@
 
 using cudf::io::io_type;
 
-enum class column_selection { ALL, ALTERNATE, HALF };
-
-enum class row_selection { ALL, BYTE_RANGE, NROWS, SKIPFOOTER };
-
 #define RD_BENCHMARK_DEFINE_ALL_SOURCES(benchmark, name, type_or_group)                  \
   benchmark(name##_file_input, type_or_group, static_cast<uint32_t>(io_type::FILEPATH)); \
   benchmark(name##_buffer_input, type_or_group, static_cast<uint32_t>(io_type::HOST_BUFFER));
@@ -77,7 +73,34 @@ class cuio_source_sink_pair {
 };
 
 /**
- * @brief Arranges columns such that total size of selected columns is half of the total table size.
+ * @brief Column selection strategy.
  */
-std::vector<cudf::type_id> opts_bm_data_types(std::vector<cudf::type_id> const& ids,
-                                              column_selection cs);
+enum class column_selection { ALL, ALTERNATE, FIRST_HALF, SECOND_HALF };
+
+/**
+ * @brief Row selection strategy.
+ *
+ * Not all strategies are applicable to all readers.
+ */
+enum class row_selection { ALL, BYTE_RANGE, NROWS, SKIPFOOTER };
+
+/**
+ * @brief Modify data types such that total selected columns size is a fix fraction of the total
+ * size.
+ *
+ * The data types are multiplied/rearranged such that the columns selected with the given column
+ * selection enumerator add up to a fixed fraction of the total table size, regardless of the data
+ * types.
+ *
+ * @param ids Array of column type IDs
+ * @param cs The column selection enumerator
+ *
+ * @return The duplicated/rearranged array of type IDs
+ */
+std::vector<cudf::type_id> dtypes_for_column_selection(std::vector<cudf::type_id> const& ids,
+                                                       column_selection cs);
+
+/**
+ * @brief Select a subset of columns based on the input enumerator.
+ */
+std::vector<int> select_columns(column_selection cs, int num_cols);
